requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.3 ; python_version >= '3.6'
aiosignal==1.3.1 ; python_version >= '3.7'
asgiref==3.6.0 ; python_version >= '3.7'
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==22.2.0 ; python_version >= '3.6'
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==2.1.1 ; python_full_version >= '3.6.0'
django==4.1.6
frozenlist==1.3.3 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
multidict==6.0.4 ; python_version >= '3.7'
openai==0.26.4
python-dotenv==0.21.1
requests==2.28.2 ; python_version >= '3.7' and python_version < '4'
sqlparse==0.4.3 ; python_version >= '3.5'
tqdm==4.64.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.14 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
yarl==1.8.2 ; python_version >= '3.7'
